requirements.txt
flask~=3.0.3
flask-cors~=4.0.1
asgiref~=3.8.1
PyYAML~=6.0.1
pillow~=10.3.0
onnxruntime~=1.18.1
numpy~=1.24.4